Refactor to clarify what is internal
