+++ RealEngine/graphics/internal_interfaces/CMakeLists.txt
@@ -1,0 +1,6 @@
+﻿target_sources(${PROJECT_NAME} PUBLIC
+	IBuffer.hpp		
+	IShaderProgram.hpp		
+	IVertexArray.hpp		
+	IViewport.hpp		
+)
