+++ RealEngine/graphics/internal_renderers/GL46_Buffer.hpp
@@ -2,7 +2,7 @@
  *  @author    Dubsky Tomas
  */
 #pragma once
-#include <RealEngine/graphics/renderers/IBuffer.hpp>
+#include <RealEngine/graphics/internal_interfaces/IBuffer.hpp>
 
 namespace RE {
