+++ RealEngine/graphics/internal_renderers/GL46_VertexArray.cpp
@@ -1,7 +1,7 @@
 ﻿/*!
  *  @author    Dubsky Tomas
  */
-#include <RealEngine/graphics/renderers/GL46_VertexArray.hpp>
+#include <RealEngine/graphics/internal_renderers/GL46_VertexArray.hpp>
 
 #include <RealEngine/graphics/VertexArray.hpp>
